Rename bfxapi.websocket.client to _client and bfxapi.websocket.handlers to _handlers (according to PEP8).
