+++ bfxapi/websocket/__init__.py
@@ -1,1 +1,1 @@
-from .client import BfxWebSocketClient, BfxWebSocketBucket, BfxWebSocketInputs
+from ._client import BfxWebSocketClient, BfxWebSocketBucket, BfxWebSocketInputs
