+++ bfxapi/websocket/_handlers/auth_events_handler.py
